Extract broker deps into interfaces.go file
